get rid of connection - move to response
